Change 'removefromclass' command to use student indices in class List.
Currently, the 'removefromclass' command uses the student indices as shown in the student list.
This is counterintuitive as the user is likely to be looking at the student list in the class tab when performing this operation.
So let us change the student indices as used in the command to used the ones shown in the class tab of the class to be removed from.